third_party/blink/web_tests/http/tests/security/powerfulFeatureRestrictions/device-orientation-on-secure-origin.https.html

<!DOCTYPE html>
<title>Device Sensor Events on Secure Origin</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/get-host-info.js"></script>
<script src="/resources/sensor-helpers.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="/gen/services/device/public/mojom/sensor_provider.mojom.js"></script>
<script>

test(() => {
  assert_true('ondevicemotion' in window);
  assert_true('ondeviceorientation' in window);
  assert_true('DeviceMotionEvent' in window);
  assert_true('DeviceOrientationEvent' in window);
  assert_true('DeviceMotionEventAcceleration' in window);
  assert_true('DeviceMotionEventRotationRate' in window);
}, 'Event handlers as well as interfaces are exposed on `window`.');

sensor_test(async (t, sensorProvider) => {
  const FAKE_ACCELERATION_DATA = [1, 2, 3];
  const FAKE_LINEAR_ACCELERATION_DATA = [4, 5, 6];
  const FAKE_GYROSCOPE_DATA = [7, 8, 9];
  setMockSensorDataForType(sensorProvider, 'Accelerometer', FAKE_ACCELERATION_DATA);
  setMockSensorDataForType(sensorProvider, 'LinearAccelerationSensor', FAKE_LINEAR_ACCELERATION_DATA);
  setMockSensorDataForType(sensorProvider, 'Gyroscope', FAKE_GYROSCOPE_DATA);

  const radToDeg = 180 / Math.PI;
  return waitForEvent(new DeviceMotionEvent('devicemotion', {
      acceleration: {
          x: FAKE_LINEAR_ACCELERATION_DATA[0],
          y: FAKE_LINEAR_ACCELERATION_DATA[1],
          z: FAKE_LINEAR_ACCELERATION_DATA[2],
      },
      accelerationIncludingGravity: {
          x: FAKE_ACCELERATION_DATA[0],
          y: FAKE_ACCELERATION_DATA[1],
          z: FAKE_ACCELERATION_DATA[2],
      },
      rotationRate: {
          alpha: FAKE_GYROSCOPE_DATA[0] * radToDeg,
          beta: FAKE_GYROSCOPE_DATA[1] * radToDeg,
          gamma: FAKE_GYROSCOPE_DATA[2] * radToDeg,
      },
      interval: 16,
  }));
}, 'DeviceMotionEvent fires.');

sensor_test(async (t, sensorProvider) => {
  const FAKE_ORIENTATION_DATA = [1.1, 2.2, 3.3];
  setMockSensorDataForType(sensorProvider, 'RelativeOrientationEulerAngles', FAKE_ORIENTATION_DATA);

  return waitForEvent(new DeviceOrientationEvent('deviceorientation', {
      alpha: FAKE_ORIENTATION_DATA[2],
      beta: FAKE_ORIENTATION_DATA[0],
      gamma: FAKE_ORIENTATION_DATA[1],
      absolute: false,
  }));
}, 'DeviceOrientationEvent fires.');

sensor_test(async (t, sensorProvider) => {
  const FAKE_ORIENTATION_DATA = [1.1, 2.2, 3.3];
  setMockSensorDataForType(sensorProvider, 'AbsoluteOrientationEulerAngles', FAKE_ORIENTATION_DATA);

  return waitForEvent(new DeviceOrientationEvent('deviceorientationabsolute', {
      alpha: FAKE_ORIENTATION_DATA[2],
      beta: FAKE_ORIENTATION_DATA[0],
      gamma: FAKE_ORIENTATION_DATA[1],
      absolute: true,
  }));
}, 'DeviceOrientationEvent fires (with absolute orientation).');
</script>
